+++ src/hooks/usePopulations.ts
@@ -1,19 +1,5 @@
 import { useState } from "react";
-
-type YearlyPopulation = {
-  year: number;
-  value: number;
-};
-type PopulationData = {
-  label: "総人口" | "年少人口" | "生産年齢人口" | "老年人口";
-  data: YearlyPopulation[];
-};
-export type PopulationCompositionPerYearAPIResponce = {
-  message: string | null;
-  result: {
-    data: PopulationData[];
-  };
-};
+import { type YearlyPopulation, resasApi } from "../utils/resasApi";
 
 type Populations = {
   total: PopulationGraphData[];
@@ -25,10 +11,6 @@ type PopulationGraphData = {
   year: number;
   [prefCode: string]: number;
 };
-
-const defaultHeaders = {
-  "X-API-KEY": import.meta.env.VITE_RESAS_API_KEY as string,
-} as const;
 
 const updatePopulations = (
   prevTotalPopulations: PopulationGraphData[],
@@ -85,39 +67,37 @@ export const usePopulations = () => {
       prefCode: prefCode.toString(),
       cityCode: "-", // 全市区町村
     });
-    const url = `https://opendata.resas-portal.go.jp/api/v1/population/composition/perYear?${queryParams}`;
-    const response = await fetch(url, {
-      headers: defaultHeaders,
-    });
-    if (!response.ok) {
-      console.error(`HTTP error! status: ${response.status} url: ${url}`);
+    try {
+      const data = await resasApi.getPopulationCompositionPerYear(
+        queryParams.toString(),
+      );
+      setPopulations((prevPopulations) => ({
+        total: updatePopulations(
+          prevPopulations.total,
+          data.result.data[0].data,
+          prefCode,
+        ),
+        young: updatePopulations(
+          prevPopulations.young,
+          data.result.data[1].data,
+          prefCode,
+        ),
+        workingAge: updatePopulations(
+          prevPopulations.workingAge,
+          data.result.data[2].data,
+          prefCode,
+        ),
+        elderly: updatePopulations(
+          prevPopulations.elderly,
+          data.result.data[3].data,
+          prefCode,
+        ),
+      }));
+    } catch {
       window.alert("データの取得に失敗しました");
       return;
     }
-    const data: PopulationCompositionPerYearAPIResponce = await response.json();
 
-    setPopulations((prevPopulations) => ({
-      total: updatePopulations(
-        prevPopulations.total,
-        data.result.data[0].data,
-        prefCode,
-      ),
-      young: updatePopulations(
-        prevPopulations.young,
-        data.result.data[1].data,
-        prefCode,
-      ),
-      workingAge: updatePopulations(
-        prevPopulations.workingAge,
-        data.result.data[2].data,
-        prefCode,
-      ),
-      elderly: updatePopulations(
-        prevPopulations.elderly,
-        data.result.data[3].data,
-        prefCode,
-      ),
-    }));
     setCheckedPrefectures((prevCheckedPrefectures) =>
       updateCheckedPrefectures(prevCheckedPrefectures, prefCode),
     );
